feat(sql-result): align standalone result panel lifecycle
